requirements_base.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.11.0
asgiref==0.13.0
autobahn==0.14.0
Babel==2.3.4
billiard==3.3.0.22
boto==2.40.0
celery==3.1.19
channels==0.13.1
colorama==0.3.7
contextlib2==0.5.3
coverage==4.0.3
daphne==0.12.0
dj-database-url==0.4.1
Django==1.8.8
django-celery==3.1.17
django-cors-headers==1.1.0
django-countries==3.4.1
django-easymoney==0.7
django-extensions==1.6.7
django-floppyforms==1.6.2
django-idmap==0.4.1
django-inspect-model==0.7
django-sslify==0.2.7
django-sslserver==0.19
django-vanilla-views==1.0.4
djangorestframework==3.3.3
funcsigs==1.0.2
honcho==0.7.1
huey==1.1.2
IPy==0.83
kombu==3.0.31
lpthw.web==1.1
mock==2.0.0
msgpack-python==0.4.7
nose==1.3.7
ordered-set==2.0.1
otree-core==0.5.5
otree-save-the-change==1.1.3
pbr==1.9.1
pytz==2016.4
raven==5.17.0
redis==2.10.5
requests==2.10.0
selenium==2.53.2
six==1.10.0
Twisted==16.1.1
txaio==2.5.1
unicodecsv==0.14.1
virtualenv==15.0.1
whitenoise==3.0
xmltodict==0.10.1
zope.interface==4.1.3
